requirements.txt
# Streamlit - Web UI
streamlit==1.31.0

# LangChain - Agent framework
langchain==0.1.9
langchain-community==0.0.24

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Vector DB
chromadb==0.4.22

# Utilities
python-dotenv==1.0.1
pandas==2.2.0